new custom hook "useCopy"
